Add more screenshots
I changed the AppStoreAssets/iPhone directory back to the AppStoreAssets/iPhone_6.7_display directory. This is because I realised that separate sizes are mandatory for iPhone 5.5 inch displays, so a new directory was needed, and I need some way of distinguishing between the two of them. All the new photos are in the AppStoreAssets/iPhone_5.5_display/ directory.

I've also included some minor changes to the README.md